pyproject.toml
[project]
name = "planning-agent"
version = "0.1.0"
description = "AI planning agent with Todoist integration and persistent context"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0",
    "pydantic>=2.0",
    "pydantic-ai>=0.1",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "requests>=2.31",
    "todoist-api-python>=3.0",
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.0",
    "itsdangerous>=2.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
]

[project.scripts]
todoist-reschedule = "todoist_scheduler.cli:main"
todoist-scheduler = "todoist_scheduler.main:main"
todoist-mcp = "todoist_mcp.server:main"
planning-context = "planning_context.server:main"
planning-agent = "planning_agent.main_cli:main_sync"
planning-agent-web = "planning_agent.main_web:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/planning_agent",
    "src/planning_context",
    "src/todoist_mcp",
    "src/todoist_scheduler",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "pyright>=1.1",
]
